old-tests/submission/W3C/domhtml/HTMLDocument15.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/domcore/#dom-document-getelementbyid" rel="help">
<link rel="author" title="NIST">
</head>
<body id="TEST-BODY"> 
<h1>This is a Document</h1>
<p>This is a paragraph.</p>
<img alt="Yellow image" height="50" src="/support/images/yellow-rectangle.png" width="100">
<p class="copyright"><a href="http://www.csail.mit.edu/"><acronym title="Massachusetts Institute of Technology">MIT</acronym></a>, <a href="http://www.ercim.eu/"><acronym title="European Research Consortium for Informatics and Mathematics">ERCIM</acronym></a>,
<a href="http://www.keio.ac.jp/" id='mapid'>Keio</a></p>


<script>
var domtest = async_test("The getElementById method returns the Element whose id is given by     elementId.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var elementNode;
      var elementValue;
      var doc;
      doc = window.document;
      elementNode = doc.getElementById("mapid");
      elementValue = elementNode.nodeName;

      assert_equals_autocase("element", elementValue,"A");
       
    domtest.done();
}));

</script>
<p>The getElementById method returns the Element whose id is given by
    elementId.  If no such element exists, returns null.  

    Retrieve the element whose id is "mapid".
    Check the value of the element.</p>
<div id="log"></div>
</body></html>
